+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/bot/SetMyDefaultAdministratorRights.kt
@@ -2,14 +2,14 @@ package dev.inmo.tgbotapi.requests.bot
 
 import dev.inmo.tgbotapi.requests.abstracts.SimpleRequest
 import dev.inmo.tgbotapi.types.*
-import dev.inmo.tgbotapi.types.chat.member.ChatAdministratorRightsImpl
+import dev.inmo.tgbotapi.types.chat.member.ChatCommonAdministratorRights
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.serializer
 
 @Serializable
 class SetMyDefaultAdministratorRights(
     @SerialName(rightsField)
-    val rights: ChatAdministratorRightsImpl,
+    val rights: ChatCommonAdministratorRights,
     @SerialName(forChannelsField)
     val forChannels: Boolean? = null
 ) : SimpleRequest<Boolean> {