+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/chat/members/PromoteChatMember.kt
@@ -1,5 +1,6 @@
 package dev.inmo.tgbotapi.requests.chat.members
 
+import dev.inmo.micro_utils.common.Warning
 import dev.inmo.tgbotapi.abstracts.types.UntilDate
 import dev.inmo.tgbotapi.requests.chat.abstracts.ChatMemberRequest
 import dev.inmo.tgbotapi.types.*
@@ -7,6 +8,7 @@ import kotlinx.serialization.*
 import kotlinx.serialization.builtins.serializer
 
 @Serializable
+@Warning("This method is too common. Use it with caution")
 data class PromoteChatMember(
     @SerialName(chatIdField)
     override val chatId: ChatIdentifier,
@@ -37,7 +39,13 @@ data class PromoteChatMember(
     @SerialName(canManageChatField)
     private val canManageChat: Boolean? = null,
     @SerialName(canManageTopicsField)
-    private val canManageTopics: Boolean? = null
+    private val canManageTopics: Boolean? = null,
+    @SerialName(canPostStoriesField)
+    private val canPostStories: Boolean? = null,
+    @SerialName(canEditStoriesField)
+    private val canEditStories: Boolean? = null,
+    @SerialName(canDeleteStoriesField)
+    private val canDeleteStories: Boolean? = null
 ) : ChatMemberRequest<Boolean>, UntilDate {
     override fun method(): String = "promoteChatMember"
     override val resultDeserializer: DeserializationStrategy<Boolean>
@@ -45,3 +53,109 @@ data class PromoteChatMember(
     override val requestSerializer: SerializationStrategy<*>
         get() = serializer()
 }
+
+fun PromoteChatMember(
+    chatId: ChatIdentifier,
+    userId: UserId,
+    untilDate: TelegramDate? = null,
+    isAnonymous: Boolean? = null,
+    canChangeInfo: Boolean? = null,
+    canDeleteMessages: Boolean? = null,
+    canInviteUsers: Boolean? = null,
+    canRestrictMembers: Boolean? = null,
+    canPromoteMembers: Boolean? = null,
+    canManageVideoChats: Boolean? = null,
+    canManageChat: Boolean? = null,
+) = PromoteChatMember(
+    chatId = chatId,
+    userId = userId,
+    untilDate = untilDate,
+    isAnonymous = isAnonymous,
+    canChangeInfo = canChangeInfo,
+    canPostMessages = null,
+    canEditMessages = null,
+    canDeleteMessages = canDeleteMessages,
+    canInviteUsers = canInviteUsers,
+    canRestrictMembers = canRestrictMembers,
+    canPinMessages = null,
+    canPromoteMembers = canPromoteMembers,
+    canManageVideoChats = canManageVideoChats,
+    canManageChat = canManageChat,
+    canManageTopics = null,
+    canPostStories = null,
+    canEditStories = null,
+    canDeleteStories = null
+)
+
+fun PromoteChannelAdministrator(
+    chatId: ChatIdentifier,
+    userId: UserId,
+    untilDate: TelegramDate? = null,
+    isAnonymous: Boolean? = null,
+    canChangeInfo: Boolean? = null,
+    canPostMessages: Boolean? = null,
+    canEditMessages: Boolean? = null,
+    canDeleteMessages: Boolean? = null,
+    canInviteUsers: Boolean? = null,
+    canRestrictMembers: Boolean? = null,
+    canPromoteMembers: Boolean? = null,
+    canManageVideoChats: Boolean? = null,
+    canManageChat: Boolean? = null,
+    canPostStories: Boolean? = null,
+    canEditStories: Boolean? = null,
+    canDeleteStories: Boolean? = null
+) = PromoteChatMember(
+    chatId = chatId,
+    userId = userId,
+    untilDate = untilDate,
+    isAnonymous = isAnonymous,
+    canChangeInfo = canChangeInfo,
+    canPostMessages = canPostMessages,
+    canEditMessages = canEditMessages,
+    canDeleteMessages = canDeleteMessages,
+    canInviteUsers = canInviteUsers,
+    canRestrictMembers = canRestrictMembers,
+    canPinMessages = null,
+    canPromoteMembers = canPromoteMembers,
+    canManageVideoChats = canManageVideoChats,
+    canManageChat = canManageChat,
+    canManageTopics = null,
+    canPostStories = canPostStories,
+    canEditStories = canEditStories,
+    canDeleteStories = canDeleteStories
+)
+
+fun PromoteSupergroupAdministrator(
+    chatId: ChatIdentifier,
+    userId: UserId,
+    untilDate: TelegramDate? = null,
+    isAnonymous: Boolean? = null,
+    canChangeInfo: Boolean? = null,
+    canDeleteMessages: Boolean? = null,
+    canInviteUsers: Boolean? = null,
+    canRestrictMembers: Boolean? = null,
+    canPinMessages: Boolean? = null,
+    canPromoteMembers: Boolean? = null,
+    canManageVideoChats: Boolean? = null,
+    canManageChat: Boolean? = null,
+    canManageTopics: Boolean? = null,
+) = PromoteChatMember(
+    chatId = chatId,
+    userId = userId,
+    untilDate = untilDate,
+    isAnonymous = isAnonymous,
+    canChangeInfo = canChangeInfo,
+    canPostMessages = null,
+    canEditMessages = null,
+    canDeleteMessages = canDeleteMessages,
+    canInviteUsers = canInviteUsers,
+    canRestrictMembers = canRestrictMembers,
+    canPinMessages = canPinMessages,
+    canPromoteMembers = canPromoteMembers,
+    canManageVideoChats = canManageVideoChats,
+    canManageChat = canManageChat,
+    canManageTopics = canManageTopics,
+    canPostStories = null,
+    canEditStories = null,
+    canDeleteStories = null
+)