+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/AdministratorChatMemberImpl.kt
@@ -35,9 +35,16 @@ data class AdministratorChatMemberImpl(
     @SerialName(customTitleField)
     override val customTitle: String? = null,
     @SerialName(canManageTopicsField)
-    override val canManageTopics: Boolean = false
+    override val canManageTopics: Boolean = false,
+    @SerialName(canPostStoriesField)
+    override val canPostStories: Boolean = false,
+    @SerialName(canEditStoriesField)
+    override val canEditStories: Boolean = false,
+    @SerialName(canDeleteStoriesField)
+    override val canDeleteStories: Boolean = false
 ) : AdministratorChatMember {
     @SerialName(statusField)
     @Required
-    private val type: String = "administrator"
+    override val status: ChatMember.Status
+        get() = ChatMember.Status.Administrator
 }