+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/ChatMember.kt
@@ -4,6 +4,7 @@ import dev.inmo.tgbotapi.abstracts.WithUser
 import dev.inmo.tgbotapi.types.statusField
 import dev.inmo.tgbotapi.utils.RiskFeature
 import dev.inmo.tgbotapi.utils.nonstrictJsonFormat
+import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.descriptors.SerialDescriptor
@@ -13,7 +14,18 @@ import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.jsonPrimitive
 
 @Serializable(ChatMemberSerializer::class)
-sealed interface ChatMember : WithUser
+sealed interface ChatMember : WithUser {
+    @Serializable
+    enum class Status(val status: String, val deserializationStrategy: DeserializationStrategy<ChatMember>) {
+        Creator("creator", OwnerChatMember.serializer()),
+        Administrator("administrator", AdministratorChatMemberImpl.serializer()),
+        Member("member", MemberChatMemberImpl.serializer()),
+        Restricted("restricted", RestrictedChatMember.serializer()),
+        Left("left", LeftChatMemberImpl.serializer()),
+        Kicked("kicked", KickedChatMember.serializer())
+    }
+    val status: Status
+}
 
 @RiskFeature
 object ChatMemberSerializer : KSerializer<ChatMember> {
@@ -21,15 +33,14 @@ object ChatMemberSerializer : KSerializer<ChatMember> {
 
     override fun deserialize(decoder: Decoder): ChatMember {
         val json = JsonObject.serializer().deserialize(decoder)
-        return when (json[statusField] ?.jsonPrimitive ?.content ?: error("Status field of chat member must be specified, but incoming json contains next: $json")) {
-            "creator" -> nonstrictJsonFormat.decodeFromJsonElement(OwnerChatMember.serializer(), json)
-            "administrator" -> nonstrictJsonFormat.decodeFromJsonElement(AdministratorChatMemberImpl.serializer(), json)
-            "member" -> nonstrictJsonFormat.decodeFromJsonElement(MemberChatMemberImpl.serializer(), json)
-            "restricted" -> nonstrictJsonFormat.decodeFromJsonElement(RestrictedChatMember.serializer(), json)
-            "left" -> nonstrictJsonFormat.decodeFromJsonElement(LeftChatMemberImpl.serializer(), json)
-            "kicked" -> nonstrictJsonFormat.decodeFromJsonElement(KickedChatMember.serializer(), json)
-            else -> error("Unknown type of chat member in json: $json")
-        }
+        val status = json[statusField] ?.jsonPrimitive ?.content ?: error("Status field of chat member must be specified, but incoming json contains next: $json")
+        return ChatMember.Status.values().firstNotNullOfOrNull {
+            if (it.status == status) {
+                nonstrictJsonFormat.decodeFromJsonElement(it.deserializationStrategy, json)
+            } else {
+                null
+            }
+        } ?: error("Unknown type of chat member in json: $json")
     }
 
     override fun serialize(encoder: Encoder, value: ChatMember) {