+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/LeftChatMemberImpl.kt
@@ -5,8 +5,12 @@ import dev.inmo.tgbotapi.types.chat.User
 import kotlinx.serialization.*
 
 @Serializable
-data class LeftChatMemberImpl(@SerialName(userField) override val user: User) : LeftChatMember {
+data class LeftChatMemberImpl(
+    @SerialName(userField)
+    override val user: User
+) : LeftChatMember {
     @SerialName(statusField)
     @Required
-    private val type: String = "left"
+    override val status: ChatMember.Status
+        get() = ChatMember.Status.Left
 }