+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/OwnerChatMember.kt
@@ -36,7 +36,15 @@ data class OwnerChatMember(
     override val canManageChat: Boolean = true
     @Transient
     override val canManageTopics: Boolean = true
+    @Transient
+    override val canPostStories: Boolean = true
+    @Transient
+    override val canEditStories: Boolean = true
+    @Transient
+    override val canDeleteStories: Boolean = true
+
     @SerialName(statusField)
     @Required
-    private val type: String = "creator"
+    override val status: ChatMember.Status
+        get() = ChatMember.Status.Creator
 }